For /plugins/editors/tinymce.xml
Change"URLS" to default="1" (which is fixed instead of relative)
This solves a lot of the problems caused by baseurl most notably when j! is installed in a subdirectory.

Change "Save Warning"  to  default="1" 
This is more like the behavior of many editors people are familiar with.


For /administrator/components/com_content/config.xml

Change linked titles to default="1"
That way a simple section or category link gives a simple linked list of articles.

Change show_category to default="1"
Change link_category to default="1"
That way a simple section link will give a list of categories that are linkable.


Change show_item_navigation to default="1"
This will generally make it easier for beginners to navigate their sites

git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@9377 6f6e1ebd-4c2b-0410-823f-f34bde69bce9