Created option for Menu Article to choose Layout option
Found an issue in the Active Menu item where there was no way to choose a HTML override option for an article. If the article was a active menu and a layout override was chosen in the article it would not work. So a option was needed in the article menu to resolve this.

New option placed in the menu area for article selection called Choose Layout.
Updated logic to check for this new field in the view for com_content/view/article/

Signed-off-by: Steven <steven@corephp.com>