fix the BUGfix the BUG http://forum.joomla.org/viewtopic.php?f=579&t=725049
Missing 'administrator' part in the form's action page of the admin login. The form's action has value '/ index.php' instead of 'administrator / index.php'.
description For my new site, I want to put Joomla in excellant own sub-directory but hide the directory name in the URL. I Followed the online instructions and everything was working well except for the admin page. The problem is That the login form's action is missing the 'administrator' part in the path. Hence I keep getting 404 error because the page tries to load from the main site's index page.fix the BUG
Missing 'administrator' part in the form's action page of the admin login. The form's action has value '/ index.php' instead of 'administrator / index.php'.


Signed-off-by: Ufbamata26 <paty1.melo@gmail.com>