Fix issues with Oracle DB Unit
This commit provides some extra functionality and fixes to the Oracle DB Unit base case. It fixes an issue with the URI parsing where the preceding slash was included with the database name which triggered errors and it also supports passing in a custom Oracle database schema and table prefix as well. A bug with the charset is also fixes to use the right field instead of the host field it was previously using.