Update modules/mod_finder/mod_finder.php
The Smart Search module was not using the selected filter in the getSearchRoutecall.It is referencing a non-existent parameter 'f' when the filter is contained inthe parameter 'searchfilter'
When a static filter is selected in the module parameters, no hidden input iscreated to specify the 'f' option. This results in the filter not being appliedduring the search.
Changed line 46 from:

$route = FinderHelperRoute::getSearchRoute($params->get('f', null));
To:
$route = FinderHelperRoute::getSearchRoute($params->get('searchfilter',null));

And the hidden input is created and the search takes the filter into accountand restricts results based on the static filter.
Note: other references to $params->get('f') were fixed in tracker item #27933 butthis one was missed.