Distinguish category descript for contentPrepare
This loop is for displaying articles in a category -- for a category blog page, for instance.   If the context isn't set to com_content.article, there is no way for the contentPrepare() plugin callback to distinguish the category description from the articles below it.   This fix is necessary to allow extension code to handle contentPrepare() callbacks correctly for category blog pages.