Update libraries/legacy/updater/adapters/extension.php
The helper function "getClientInfo" takes two arguments, the search "$id" and "$byName" which decides if it should search by id or name.
In our case here we want to search by the name and thus it should be set to "1". Otherwise the function will return null and break all extension updates as they are not found.