Update modules/mod_articles_archive/helper.php
Fixed the problem described in this this ticket - http://joomlacode.org/gf/project/joomla/tracker/?action=TrackerItemEdit&tracker_item_id=29666

While doing so I found out, that the listing of dates of archived articles in the module is done by "Created date", but the listing of archived articles in the view is done by "Start Publishing". Which one is the correct one? Both of them should be the same for the sake of consistency.