Fix and expand JDatabaseQuery tests.
Fixed mocking of the quote method in TestMockDatabaseDriver to allow for
an array of strings to be quoted.
Fixed a bug in JDatabaseQuery->quote that escaped strings before sending
them to the JDatabaseDriver->quote method ($escape argument was not
being honoured).
Added the escape method to TestMockDatabaseDriver.
Added covers to and improved code coverage of the JDatabaseQuery class.