Added a update mechanism for the Update component
See:
https://groups.google.com/d/msg/joomla-dev-cms/2iuCF8bsA44/Nkec8TAKFV0J

The above code would allow us to have a fallback option to update the Joomla! Update component in case something goes drastically wrong with it.  Currently, there is not an xml at ttp://update.joomla.org/core/extensions/joomla/com_joomlaupdate.xml , but it's not needed since this would only be used on an as needed basis. 