Removed colors from Icomoon named icons
Only the Joomla (JUI) icons classes will get the assigned colors. This way you can still use the original Icomoon classnames to get the icon without the color.
Example:
- icon-apply and icon-edit will be blue
- icon-pencil will not be specifically colored

This change should not affect anything in the current Isis template (as it uses the JUI class names, and their styles remain the same). It does however give developers the ability to use the icons without the assigned colors.