Added (min) height to table cells with context
Fixes difference in height between normal and hover on cell with context.

The context (button with submenu) only appears on hover. This however makes the cell a little higher (when the row only has 1 line content).
This causes a weird 'jitter' effect, which you can see in the module manager for instance.

This patch fixes that by giving the cell a (min) height, making the non-hover cell high enough to fit the context menu.