JTable::getInstance quits too easily
This one we may want to talk about..there may be a better place and a better way to fix it. What happens is if you have two files that fit the name, but the first one doesn't have the right class, JTable quits without ever seeing the second one. I can see a discussion, though, about whether we want to enforce good naming conventions. It's an ideal vs real discussion.