Remove &moduleclass_sfx from <input> and <button> elements
Removing the module class suffix from <input> and <button> elements in mod_search. They're not needed for anything but it may well break layouts if a suffix is entered. Also it's the only module where it's applied to those elements. Normally it's only applied to the parent <div> where it makes sense.