Fix calender popup on enter
[#30413] Calender popup problem in pages containing calendar fields
http://joomlacode.org/gf/project/joomla/tracker/?action=TrackerItemEdit&tracker_item_id=30413

When we press enter while a text field is in focus browser automatically submits the form. So to submit a form browser triggers click event on the first submit button present in dom hierarchy. So the first button is calender display button, because of that calender popup is coming. 

There is no need to be calender display button to be submit button. So to make it of type="button" instead of type="submit" (by default id nothing is specified in button tag then it is a submit button) I added type="button" in button tag.