Reflect default values for weblink icons in layout
In the com_weblinks component settings, the default value for the "icons" is 1 (show icons), which is also states by the tooltip. However in the layoutfile the default value was zero (show text). This produced an unexpected behaviour as long as the settings were not saved.
This commit fixes this default value, showing the icon by default even when the settings aren't saved yet.
Also the check for "link_icons" was redundant if the parameter wasn't set. There is no need to check for something we already know isn't set :-)