FIX: truncating a string fail when last space is within a tag
when you have a string "..... this is text
Here is the limit" and the limit is within "Here". The next space is the space between
(there could also be any other tag with spaces) what is wrong because we end up with not valid html code. The right place is the space between "is" and "text".

This fix is checking if the place is within a tag and moves it before the tag.