Fix issue of the formvalidation not loading the Mootols More
The validate.js still uses mootools stuff (and jQuery too!) and relies on the mootools-more script.
However, the mootools-more is not loaded by the formvalidation method. This was getting loaded before, because of the mootools tooltips. But the tooltips now use the Bootstrap tooltips. So now the validation is broken.

So the fix is to make the formvalidation method load the mootools-more script (and not the mootools-core).
That's what this PR does.

The best solution is of course to remove the mootools stuff out of that script entirely and convert it to jQuery.

#### Testing instructions
Go to a user edit page (index.php?option=com_users&view=user&layout=edit&id=...) and try to save it. Before fix it will fail (nothing happens except javascript error in console about Hash). After it should work.