Offline template: remove "remember me" checkbox when plugin disabled
See following bug report: http://joomlacode.org/gf/project/joomla/tracker/?action=TrackerItemEdit&tracker_item_id=31716

Note: This is a take-over of the check for activation of the "remember" plugin from in \templates\atomic\html\mod_login\default.php on Joomla 2.5.14.

I tested it on Joomla 1.5, but can't test it on the current version, because I don't use it. While it seems from the code that the same (names) are still used in the current version, it still needs to be tested. I'm going to add a test procedure to the above bug report, for someone to test this quickly.