Fixing the Postinstall message in cpanel
`FOFModel::getTmpInstance` does reset the input before the actualy query is run. Thus the `eid` passed in the `input` array is not recognized.
It works in cPanel because it defaults to 700. If you change the eid number to some random value, you will see that the message still shows. But it shouldn't then.

Changing the call to `FOFModel::getAnInstance()` solves this.

It's possible that there are better fixes. It's just the first one I came up with.

This isn't a real issue in cPanel itself. Because of the intern default value it still works. I discovered it because I wanted to use the code to get messages for com_installer.