Improve $spans calculation in cPanel layout
The cPanel layout calculates the `spans` to start a new line if it goes over 12. This works fine as long as `bootstrapSize` not 0. If one or more module is set to 0, the design goes crazy because it doesn't start a new line, but the module is shown with full width.

This PR sets the bootstrapSize to 12 for the calculation so a new line is properly created.