Update user.php
The method for changing the password into the database should be the same if the user is new or if the user changes his profile. An other issue is for the getCryptedPassword function. It gives the salt value in the result if the encryption is sha256 or if the encryption is bcrypt (default value) and the use of Strong Encryption is set.