Fix for running validation on disabled fields which where dynamically created
Fixes an issue with fields that have been created dynamically. Prior to this fix, it runs validation even on disabled fields. 

In my case this caused trouble with twitter's typeahead.js autocomplete. Though the "tt-hint" input field was disabled, it would be validated, which caused the submit button having to be pressed twice in order to submit the form.