[PR] Let php handle the default session storage
It seems that used to be the way to handle the default php session handler in the platform until 12.2 and this commit changed it:
https://github.com/joomla/joomla-platform/commit/a47f429c50852bf2a7a36e457639b6c414ded16e

I think the choice for the default handler should be left to php.
For instance Redis is not supported in the CMS at this time but if it might be configured in php.ini to be used as default for some reason.

Please see https://groups.google.com/forum/#!topic/joomla-dev-general/stbltpP5QX4 for an example.

This also blocks installing Joomla on read-only filesystems (eg. Google AppEngine) since the 'file' handler does not work on it and memcache is the default and ONLY choice.

Freamwork doesn't force the handler either.