[imp] Use alternative CURL when allow_url_open is OFF. Fixes #2807
We propose to add an alternative method to get the feed when the allow_url_open is OFF.
When the allow_url_open is OFF, the XMLReader::open always fails because the URL can not be connected.

So we propose to get the Feed with the alternative CURL when it is present.

New proposal to use CURL when the allow_url_open is OFF

Based on the comment, here it is a new proposal that use the JHttpFactory::getHttp() function instead of the CURL nativelly.

Update factory.php

Update factory.php