Stop config html.view.templates details cascading to future config html.view.templates
The ConfigViewCmsHtml is just overwritten for each Loadtemplate() call...any parameters NOT overwritten by a future template will perpetuate.  IE add a description into the site (site settings) configuration template.  It will be written to all subsequent templates.

This patch provides a quick fix that resets the name, fieldname and description after the output buffer is cleared and the template  _output is returned.