Create session.php
This is the counterpart to the system/session plugin. Here the backend changes to a user object are tracked and a 'refresh' flag is added to the related user's session, which then triggers the user's session to be silently updated in the system/session plugin withouth the user to log out and in again.