removed type hinting
The reason for doing this is that it breaks FOF tag implementation, while this only affects a couple of files I removed it for JTable* classes in the whole repository so that we don't use Type Hinting for JTable. It may be a good idea but it doesn't work in the real world with developing extensions for Joomla!