Update intro_image.php
Reference: https://groups.google.com/forum/#!topic/joomla-dev-cms/2l8Xy0MaAfo
Standard behaviour in blogs is that images are linked to blog item full text page. Currently only the title can be linked to the full text. The proposed change adds a link to blog image.
Disadvantage: there is no parameter to switch the link off. So if a user does not want the images linked, the layout needs to be overridden. Currently users who DO want the link need to override.
I believe the majority of users will want the images linked, so unlike now, in the future only a minority will need to create an override.
Ideal would be an additional parameter next to the setting for linked titles.