Make testSortObjects() compatible with HHVM
In Zend PHP, order in the result output of usort() function is undefined, if two items has the same key value. Therefore, this means that JArrayHelper::sortObjects can return two items with the same key in any order as it uses usort() in its implementation. However, testSortObjects() assumes that usort() is always consistent, which is not guaranteed and at least not in the case of HHVM.

This patch will make testSortObjects() works in both Zend and HHVM PHP implementations.