Update TagManager0001Test.php
on line 117 the command was checking whether the tag was in row two but instead of that it was checking that if it was present in row 1.
on line 137 it was checking whether tag that was added its row number should be one but that's not necessary as if there exited some tags already then its row number would be accordingly as it should be but not 1. so now I am just checking here that its row number is greater or equal to 1.