For issue #33663
plgContentLoadmodule::_loadmod() caches content it creates using just the module type as the index.  If more than one module of the same type is included using {loadmodule}, whichever module renders first provides the content for all instances of that module type.

The fix uses a concatenation of the module type and title, providing a unique index.