Avoid "duplicate key on insert detected" in method addContentType
When indexing content via a finder plugin which may have different mime types (e.g. document management systems), adding a new mime will work, but if another document with this mime type exists, consecutive getting the type id results in a duplicate insert in table #__finder_types.
This is due to that the static method variable $types is initialized on the first call of the method.

The fix adds the newly inserted mime type in the $types hash, so that calling this method againg with the same mimetype results in a positive lookup within the hash.