Add fast check for image orientation
Detecting an image's orientation (landscape, portrait, etc) is a repetitive task when is comes to decide which values to set for width and height to the resize() method. I found my self in need to know if the images i process are landscape or portrait or rather rare in cubic format. As a first step i coded this check where i needed it which i found pretty unhandy. Querying this information statically like is possible with the getImageFileProperties() method was a very helpful fast accessor for this information.

At another stage (when preparing my view) i am again required to know an image's orientation to apply a CSS class. As can be seen the need for this information is there which is why i add this pull request including the 3 new methods: isCubic(), isLandscape(), isPortrait().

Any feedback is welcome.