User profile website only cleaned, when DOB is set. Fixes #3707
Within the onUserBeforeSave() event, the user profile data are checked to make sure they are valid. The "website" value is converted into Puny-code, while the DOB (Date-Of-Birth) is converted from whatever-format to MySQL-format. However, the "website" value is only checked for, when the DOB value is set. The patch places the "website" check outside of the "dob" check, and into its own if-structure.

To test this, I've used a workaround to see whether the PHP-line itself was executed. I tried to enter some kind of non-Punycode URL that would be converted into proper Punycode but I have failed with this so far. If somebody else has suggestions for this (so: how to enter a non-Punycode URL for testing purpose) that would be awesome.

To test the code with my hack, just add the following line, right after the line mentioning the Puny-code:
$data['profile']['website'] = 'http://example.com';

Test without patch:
* Enable the profile plugin so you are able to enter both website as DOB in your profile.
* Hack the code as mentioned above
* Enter no DOB, but do add a website URL (not "http://example.com").
* The website is saved properly.
* Repeat this, but now enter a DOB and a website URL.
* The hack should change the URL to "http://example.com".

The test above should show you that the hack of setting the URL to "example.com" is only applied when the DOB is set, and it is not applied when the DOB is not set.

Test with patch:
* Enable the profile plugin so you are able to enter both website as DOB in your profile.
* Hack the code as mentioned above
* Enter no DOB, but do add a website URL (not "http://example.com").
* The hack should change the URL to "http://example.com".
* Repeat this, but now enter a DOB and a website URL.
* The hack should change the URL to "http://example.com".

The test with patch should result into the hack being applied all the time - so changing the URL to "example.com", regardless of whether the DOB is set or not.

The test can be run from either the backend (for any user profile) or the frontend (for only your own profile).

Update profile.php

Remove check for Puny URL entirely, because it is already being filtered through the "url" filter. Removed some unneeded white lines as well.